CQ-4344327 rename variables, moving shortdescription to field
