+++ bundles/core/src/main/java/com/adobe/cq/forms/core/components/models/form/Field.java
@@ -15,7 +15,10 @@
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
 package com.adobe.cq.forms.core.components.models.form;
 
+import org.jetbrains.annotations.Nullable;
 import org.osgi.annotation.versioning.ConsumerType;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
 
 /**
  * Defines the form {@code Field} Sling Model used for form field component.
@@ -87,4 +90,26 @@ public interface Field extends Base, BaseConstraint {
      */
     Object getDefault();
 
+    /**
+     * Returns the short description of the field
+     *
+     * @return the short description of the field
+     * @since com.adobe.cq.forms.core.components.models.form 1.0.0
+     */
+    @JsonIgnore
+    @Nullable
+    default String getShortDescription() {
+        return null;
+    }
+
+    /**
+     * Returns {@code true} if short description should always be visible, otherwise {@code false}.
+     *
+     * @return {@code true} if short description should always be visible, otherwise {@code false}
+     * @since com.adobe.cq.forms.core.components.models.form 1.0.0
+     */
+    @JsonIgnore
+    default boolean isShortDescriptionVisible() {
+        return false;
+    }
 }
