+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/textinput/v1/textinput/textinput.html
@@ -14,12 +14,12 @@
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
 
-<sly data-sly-use.logic="${'textinput.js'}"
+<sly data-sly-use.renderer="${'textinput.js'}"
      data-sly-use.clientlib="${'/libs/granite/sightly/templates/clientlib.html'}"
-     data-sly-use.label="${logic.labelPath}"
-     data-sly-use.shortDescription="${logic.shortDescriptionPath}"
-     data-sly-use.longDescription="${logic.longDescriptionPath}"
-     data-sly-use.questionMark="${logic.questionMarkPath}"></sly>
+     data-sly-use.label="${renderer.labelPath}"
+     data-sly-use.shortDescription="${renderer.shortDescriptionPath}"
+     data-sly-use.longDescription="${renderer.longDescriptionPath}"
+     data-sly-use.questionMark="${renderer.questionMarkPath}"></sly>
 <div class="cmp-adaptiveform-textinput"
      data-sly-use.text="com.adobe.cq.forms.core.components.models.form.TextInput"
      data-sly-use.component="com.adobe.cq.wcm.core.components.models.Component">
@@ -48,5 +48,5 @@
     <div data-sly-call="${shortDescription.shortDescription @componentId=component.id, shortDescriptionVisible=text.shortDescriptionVisible, shortDescription=text.shortDescription, bemBlock='cmp-adaptiveform-textinput'}">
     </div>
     <div data-sly-call="${longDescription.longDescription @componentId=component.id, longDescription=text.description, bemBlock='cmp-adaptiveform-textinput'}"></div>
-    <sly data-sly-call="${clientlib.all @ categories=logic.clientlibs}"></sly>
+    <sly data-sly-call="${clientlib.all @ categories=renderer.clientlibs}"></sly>
 </div>
