+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/textinput/v1/textinput/textinput.js
@@ -17,10 +17,10 @@
 use(function () {
 
   var clientlibsArr = ['core.forms.components.base.v1.editor'];
-  var labelPath = 'core/fd/components/commons/v1/formTemplates/label.html';
-  var shortDescriptionPath = "core/fd/components/commons/v1/formTemplates/shortDescription.html";
-  var longDescriptionPath = "core/fd/components/commons/v1/formTemplates/longDescription.html";
-  var questionMarkPath = "core/fd/components/commons/v1/formTemplates/questionMark.html"
+  var labelPath = 'core/fd/components/commons/v1/fieldTemplates/label.html';
+  var shortDescriptionPath = "core/fd/components/commons/v1/fieldTemplates/shortDescription.html";
+  var longDescriptionPath = "core/fd/components/commons/v1/fieldTemplates/longDescription.html";
+  var questionMarkPath = "core/fd/components/commons/v1/fieldTemplates/questionMark.html"
   return {
     labelPath: labelPath,
     shortDescriptionPath: shortDescriptionPath,
